requirements.txt
google-generativeai
python-dotenv
openai
gradio
langchain
langchain-weaviate